Disk IO library: fix integer overflow in IO buffer pool get method with buffer pool larger than 4GB.

+++ src/cc/qcdio/QCIoBufferPool.cc
@@ -136,7 +136,7 @@ public:
             theNext <= BufferIndex(mTotalCnt));
         *mFreeListPtr = theNext;
         mFreeListPtr[theIdx] = kInUse;
-        return (mStartPtr + ((theIdx - 1) << mBufSizeShift));
+        return (mStartPtr + (size_t(theIdx - 1) << mBufSizeShift));
     }
 
     bool Put(char* inPtr)
